% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primus_projects_functions.R
\name{primus_model_project_create}
\alias{primus_model_project_create}
\title{Primus Machine Learning Project Structure Function}
\usage{
primus_model_project_create(
  folders = c("00_data", "01_data_exp", "02_data_prep", "03_models", "04_model_eval",
    "05_reports", "06_scripts", "07_visualizations")
)
}
\arguments{
\item{folders}{input folders to create. Default data, data exp, data prep, models, model_eval, reports, scripts, visualizations}
}
\value{
No return, just creates projects folders
}
\description{
Primus Machine Learning Project Structure Function
}
\examples{

require(fs)
primus_model_project_create()

}
