% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primus_projects_functions.R
\name{primus_model_project_delete}
\alias{primus_model_project_delete}
\title{Title}
\usage{
primus_model_project_delete(
  folders = c("00_data", "01_data_exp", "02_data_prep", "03_models", "04_model_eval",
    "05_reports", "06_scripts", "07_visualizations")
)
}
\arguments{
\item{folders}{input folders to remove}
}
\value{

}
\description{
Title
}
\examples{

require(fs)
primus_model_project_delete

}
